Split mongo into separate classes
