+++ packages/database-mongo/src/mongo.ts
@@ -1,23 +1,13 @@
-import {
-  ConnectionInformations,
-  CreateUser,
-  DatabaseInterface,
-  Session,
-  User,
-} from '@accounts/types';
-import { get, merge } from 'lodash';
+import { CreateUser, DatabaseInterface, User } from '@accounts/types';
+import { merge } from 'lodash';
 import { Collection, Db, ObjectID } from 'mongodb';
 
 import { AccountsMongoOptions, MongoUser } from './types';
+import { MongoSessions } from './sessions';
+import { MongoPassword } from './services/password';
+import { toMongoID } from './utils';
 
-const toMongoID = (objectId: string | ObjectID) => {
-  if (typeof objectId === 'string') {
-    return new ObjectID(objectId);
-  }
-  return objectId;
-};
-
-const defaultOptions = {
+export const defaultOptions = {
   collectionName: 'users',
   sessionCollectionName: 'sessions',
   timestamps: {
@@ -37,8 +27,9 @@ export class Mongo implements DatabaseInterface {
   private db: Db;
   // Account collection
   private collection: Collection;
-  // Session collection
-  private sessionCollection: Collection;
+
+  public sessions: MongoSessions;
+  public password: MongoPassword;
 
   constructor(db: any, options?: AccountsMongoOptions) {
     this.options = merge({ ...defaultOptions }, options);
@@ -47,22 +38,13 @@ export class Mongo implements DatabaseInterface {
     }
     this.db = db;
     this.collection = this.db.collection(this.options.collectionName);
-    this.sessionCollection = this.db.collection(this.options.sessionCollectionName);
+    this.sessions = new MongoSessions(this.db, this.options);
+    this.password = new MongoPassword(this.db, this.options);
   }
 
   public async setupIndexes(): Promise<void> {
-    await this.sessionCollection.createIndex('token', {
-      unique: true,
-      sparse: true,
-    });
-    await this.collection.createIndex('username', {
-      unique: true,
-      sparse: true,
-    });
-    await this.collection.createIndex('emails.address', {
-      unique: true,
-      sparse: true,
-    });
+    await this.sessions.setupIndexes();
+    await this.password.setupIndexes();
   }
 
   public async createUser({
@@ -125,34 +107,6 @@ export class Mongo implements DatabaseInterface {
     return user;
   }
 
-  public async findPasswordHash(userId: string): Promise<string | null> {
-    const user = await this.findUserById(userId);
-    if (user) {
-      return get(user, 'services.password.bcrypt');
-    }
-    return null;
-  }
-
-  public async findUserByEmailVerificationToken(token: string): Promise<User | null> {
-    const user = await this.collection.findOne({
-      'services.email.verificationTokens.token': token,
-    });
-    if (user) {
-      user.id = user._id.toString();
-    }
-    return user;
-  }
-
-  public async findUserByResetPasswordToken(token: string): Promise<User | null> {
-    const user = await this.collection.findOne({
-      'services.password.reset.token': token,
-    });
-    if (user) {
-      user.id = user._id.toString();
-    }
-    return user;
-  }
-
   public async findUserByServiceId(serviceName: string, serviceId: string): Promise<User | null> {
     const user = await this.collection.findOne({
       [`services.${serviceName}.id`]: serviceId,
@@ -163,60 +117,20 @@ export class Mongo implements DatabaseInterface {
     return user;
   }
 
-  public async addEmail(userId: string, newEmail: string, verified: boolean): Promise<void> {
+  public async setUserDeactivated(userId: string, deactivated: boolean): Promise<void> {
     const id = this.options.convertUserIdToMongoObjectId ? toMongoID(userId) : userId;
-    const ret = await this.collection.updateOne(
+    await this.collection.updateOne(
       { _id: id },
       {
-        $addToSet: {
-          emails: {
-            address: newEmail.toLowerCase(),
-            verified,
-          },
-        },
         $set: {
+          deactivated,
           [this.options.timestamps.updatedAt]: this.options.dateProvider(),
         },
       }
     );
-    if (ret.result.nModified === 0) {
-      throw new Error('User not found');
-    }
   }
 
-  public async removeEmail(userId: string, email: string): Promise<void> {
-    const id = this.options.convertUserIdToMongoObjectId ? toMongoID(userId) : userId;
-    const ret = await this.collection.updateOne(
-      { _id: id },
-      {
-        $pull: { emails: { address: email.toLowerCase() } },
-        $set: {
-          [this.options.timestamps.updatedAt]: this.options.dateProvider(),
-        },
-      }
-    );
-    if (ret.result.nModified === 0) {
-      throw new Error('User not found');
-    }
-  }
-
-  public async verifyEmail(userId: string, email: string): Promise<void> {
-    const id = this.options.convertUserIdToMongoObjectId ? toMongoID(userId) : userId;
-    const ret = await this.collection.updateOne(
-      { _id: id, 'emails.address': email },
-      {
-        $set: {
-          'emails.$.verified': true,
-          [this.options.timestamps.updatedAt]: this.options.dateProvider(),
-        },
-        $pull: { 'services.email.verificationTokens': { address: email } },
-      }
-    );
-    if (ret.result.nModified === 0) {
-      throw new Error('User not found');
-    }
-  }
-
+  // TODO this should be in the password package
   public async setUsername(userId: string, newUsername: string): Promise<void> {
     const id = this.options.convertUserIdToMongoObjectId ? toMongoID(userId) : userId;
     const ret = await this.collection.updateOne(
@@ -225,25 +139,6 @@ export class Mongo implements DatabaseInterface {
         $set: {
           username: newUsername,
           [this.options.timestamps.updatedAt]: this.options.dateProvider(),
-        },
-      }
-    );
-    if (ret.result.nModified === 0) {
-      throw new Error('User not found');
-    }
-  }
-
-  public async setPassword(userId: string, newPassword: string): Promise<void> {
-    const id = this.options.convertUserIdToMongoObjectId ? toMongoID(userId) : userId;
-    const ret = await this.collection.updateOne(
-      { _id: id },
-      {
-        $set: {
-          'services.password.bcrypt': newPassword,
-          [this.options.timestamps.updatedAt]: this.options.dateProvider(),
-        },
-        $unset: {
-          'services.password.reset': '',
         },
       }
     );
@@ -278,152 +173,5 @@ export class Mongo implements DatabaseInterface {
         },
       }
     );
-  }
-
-  public async setUserDeactivated(userId: string, deactivated: boolean): Promise<void> {
-    const id = this.options.convertUserIdToMongoObjectId ? toMongoID(userId) : userId;
-    await this.collection.updateOne(
-      { _id: id },
-      {
-        $set: {
-          deactivated,
-          [this.options.timestamps.updatedAt]: this.options.dateProvider(),
-        },
-      }
-    );
-  }
-
-  public async createSession(
-    userId: string,
-    token: string,
-    connection: ConnectionInformations = {},
-    extraData?: object
-  ): Promise<string> {
-    const session = {
-      userId,
-      token,
-      userAgent: connection.userAgent,
-      ip: connection.ip,
-      extraData,
-      valid: true,
-      [this.options.timestamps.createdAt]: this.options.dateProvider(),
-      [this.options.timestamps.updatedAt]: this.options.dateProvider(),
-    };
-
-    if (this.options.idProvider) {
-      session._id = this.options.idProvider();
-    }
-
-    const ret = await this.sessionCollection.insertOne(session);
-    return (ret.ops[0]._id as ObjectID).toString();
-  }
-
-  public async updateSession(
-    sessionId: string,
-    connection: ConnectionInformations,
-    newToken?: string
-  ): Promise<void> {
-    const updateClause = {
-      $set: {
-        userAgent: connection.userAgent,
-        ip: connection.ip,
-        [this.options.timestamps.updatedAt]: this.options.dateProvider(),
-      },
-    };
-
-    if (newToken) {
-      updateClause.$set.token = newToken;
-    }
-
-    const _id = this.options.convertSessionIdToMongoObjectId ? toMongoID(sessionId) : sessionId;
-    await this.sessionCollection.updateOne({ _id }, updateClause);
-  }
-
-  public async invalidateSession(sessionId: string): Promise<void> {
-    const _id = this.options.convertSessionIdToMongoObjectId ? toMongoID(sessionId) : sessionId;
-    await this.sessionCollection.updateOne(
-      { _id },
-      {
-        $set: {
-          valid: false,
-          [this.options.timestamps.updatedAt]: this.options.dateProvider(),
-        },
-      }
-    );
-  }
-
-  public async invalidateAllSessions(userId: string): Promise<void> {
-    await this.sessionCollection.updateMany(
-      { userId },
-      {
-        $set: {
-          valid: false,
-          [this.options.timestamps.updatedAt]: this.options.dateProvider(),
-        },
-      }
-    );
-  }
-
-  public async findSessionByToken(token: string): Promise<Session | null> {
-    const session = await this.sessionCollection.findOne({ token });
-    if (session) {
-      session.id = session._id.toString();
-    }
-    return session;
-  }
-
-  public async findSessionById(sessionId: string): Promise<Session | null> {
-    const _id = this.options.convertSessionIdToMongoObjectId ? toMongoID(sessionId) : sessionId;
-    const session = await this.sessionCollection.findOne({ _id });
-    if (session) {
-      session.id = session._id.toString();
-    }
-    return session;
-  }
-
-  public async addEmailVerificationToken(
-    userId: string,
-    email: string,
-    token: string
-  ): Promise<void> {
-    const _id = this.options.convertUserIdToMongoObjectId ? toMongoID(userId) : userId;
-    await this.collection.updateOne(
-      { _id },
-      {
-        $push: {
-          'services.email.verificationTokens': {
-            token,
-            address: email.toLowerCase(),
-            when: this.options.dateProvider(),
-          },
-        },
-      }
-    );
-  }
-
-  public async addResetPasswordToken(
-    userId: string,
-    email: string,
-    token: string,
-    reason: string
-  ): Promise<void> {
-    const _id = this.options.convertUserIdToMongoObjectId ? toMongoID(userId) : userId;
-    await this.collection.updateOne(
-      { _id },
-      {
-        $push: {
-          'services.password.reset': {
-            token,
-            address: email.toLowerCase(),
-            when: this.options.dateProvider(),
-            reason,
-          },
-        },
-      }
-    );
-  }
-
-  public async setResetPassword(userId: string, email: string, newPassword: string): Promise<void> {
-    await this.setPassword(userId, newPassword);
   }
 }
